reverted spi_basics
