+++ lib/display_driver.cpp
@@ -33,6 +33,21 @@ inline void set_cs_low()
 	PORTB &= ~(1 << TFT_CS_PORT);
 }
 
+static inline void spi_write(uint8_t data)
+{
+	SPDR = data;
+	// Wait until write is finished
+	asm volatile("nop");
+	while (!(SPSR & _BV(SPIF)))
+		;
+}
+
+static inline void spi_write16(uint16_t data)
+{
+	spi_write(data >> 8);
+	spi_write(data);
+}
+
 static inline void display_begin_write()
 {
 	set_cs_low();
