+++ lib/display_driver.h
@@ -3,7 +3,7 @@
 #define __ILI3941_DISPLAY_DRIVER__
 
 #include <avr/delay.h>
-#include "spi_basics.h"
+#include <avr/io.h>
 
 #define TFT_DC 9
 #define TFT_CS 10
@@ -48,5 +48,7 @@ void display_begin_write();
 void display_end_write();
 void display_send_command(uint8_t command);
 void init_display();
+void spi_write(uint8_t data);
+void spi_write16(uint16_t data);
 
 #endif
