--- lib/spi_basics.h
@@ -1,11 +1,0 @@
-
-#ifndef __SPI_BASICS__
-#define __SPI_BASICS__
-
-#include <avr/io.h>
-
-void spi_write(uint8_t data);
-void spi_write16(uint16_t data);
-
-#endif
-
